Config file that stores all global settup information is added. Nova_mimic class was remastered according to new changes to logic. Tests was supplied for that new remastered code.

+++ utils/db_fixture.py
@@ -9,22 +9,33 @@ Also another main function of this module is wiping out data from
 previous launches of system. This means dropping tables from
 sql database and drop whole database which contains images from
 key-value storage.
+
+Main convinience of this module is that it can provide initial data set
+not only for production database but also fixture data for test database
+since initial state of mentioned is same as that in prodaction db.
+
+This process (processing production db or test db data) is controlled by
+argument for dump_data function. This arg has default value set to devel
+deployment argument for production depoyment. If this code is using for
+depoying test database function dump_data is called with deployment_arg
+set to testing.
 '''
 
+import sys
 import os.path
 
 import gridfs
 import pymongo
 
-from sqlalchemy import create_engine, Column, Integer, String, ForeignKey, Boolean
+from sqlalchemy import (
+    create_engine, Column, Integer,
+    String, ForeignKey, Boolean
+)
+
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import sessionmaker, relationship
 
-connection_string = 'mysql+mysqldb://devel:devel@localhost/devel?charset=utf8'
-engine = create_engine(connection_string)
-
 Base = declarative_base()
-Session = sessionmaker(bind=engine)
 
 
 class Flavors(Base):
@@ -39,11 +50,10 @@ class Flavors(Base):
 class Images(Base):
     __tablename__ = 'images'
 
-    id = Column(Integer, primary_key=True)
+    id = Column(String(50), primary_key=True, autoincrement=False)
     name = Column(String(50))
     fmt = Column(String(20))
     size = Column(Integer)
-    storage_key = Column(String(50))
 
 
 class MacAddressPool(Base):
@@ -64,19 +74,25 @@ class Instance(Base):
     mac_addr = Column(Integer, ForeignKey('mac_address_pool.id'))
 
 
-if __name__ == '__main__':
+def dump_data(deployment_arg='devel'):
+    #create proper accredetation for database
+    connection_string = 'mysql+mysqldb://{0}:{0}@localhost/{0}?charset=utf8'.format(deployment_arg)
+    engine = create_engine(connection_string)
+
     #get instance of Database object via proxy mongo client object
     #(which is listening on localhost
     #and 27017 port). This db object will be passed to GridFS instance which
     #provides toolkit for uploading files to mongoDB.
-    #Perforimg cleaning (simply dropping data base which contains images)
+    #If debug_arg is equal to 'devel' perforimg cleaning
+    #(simply dropping data base which contains images)
     #if there exists any data. Only after that one can upload image to gridfs.
     with pymongo.MongoClient() as client:
         #test presence of databases
-        if 'image_storage' in client.database_names():
-            client.drop_database('image_storage')
+        if deployment_arg == 'devel':
+            if deployment_arg in client.database_names():
+                client.drop_database(deployment_arg)
 
-        db = client.image_storage
+        db = client[deployment_arg]
         grfs = gridfs.GridFS(db)
 
         # path to image which will be uploaded to mongo storage.
@@ -92,13 +108,15 @@ if __name__ == '__main__':
         with open(path_to_image) as f:
             image_id = grfs.put(f, name="pattern_for_lv_wrapper.img")
 
-    #processing sql data: cleaning existing (if any) data
+    #processing sql data: if script is executing in production deployment
+    #cleaning existing (if any) data
     #from previous launches of system and then creating new
     #schema with filling it with needed values.
 
     #TODO: make test of schema existing more sophisticated
-    if any([engine.has_table(key) for key in Base.metadata.tables.keys()]):
-        Base.metadata.drop_all(engine)
+    if deployment_arg == 'devel':
+        if any([engine.has_table(key) for key in Base.metadata.tables.keys()]):
+            Base.metadata.drop_all(engine)
 
     Base.metadata.create_all(engine)
 
@@ -111,10 +129,10 @@ if __name__ == '__main__':
     images = []
     images.append(
         Images(
+            id=str(image_id),
             name='ubuntu12.04server',
             fmt='qcow2',
-            size=1400,
-            storage_key=str(image_id)
+            size=1400
         )
     )
 
@@ -125,6 +143,7 @@ if __name__ == '__main__':
     mac_addresses.append(MacAddressPool(address="52:54:00:83:df:a4", is_free=True))
 
     # Put objects into session and flush to database
+    Session = sessionmaker(bind=engine)
     s = Session()
 
     s.add_all(flavors)
@@ -134,3 +153,5 @@ if __name__ == '__main__':
     s.flush()
     s.commit()
     s.close()
+
+    #engine must be returne
